+++ plugin/src/main/kotlin/io/ktor/plugin/features/FatJar.kt
@@ -2,7 +2,9 @@ package io.ktor.plugin.features
 
 import com.github.jengelman.gradle.plugins.shadow.ShadowPlugin
 import com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar
+import io.ktor.plugin.internal.*
 import org.gradle.api.Project
+import org.gradle.api.plugins.ApplicationPlugin
 import org.gradle.api.tasks.TaskProvider
 
 abstract class FatJarExtension(project: Project) {
@@ -25,11 +27,23 @@ private const val RUN_FAT_JAR_TASK_DESCRIPTION =
 private const val SHADOW_RUN_TASK_NAME = "runShadow"
 private const val SHADOW_JAR_TASK_NAME = "shadowJar"
 
-fun configureFatJar(project: Project) {
-    project.plugins.apply(ShadowPlugin::class.java)
-    val tasks = project.tasks
+fun configureFatJar(project: Project) = with(project) {
+    val fatJarExtension = createKtorExtension<FatJarExtension>(FAT_JAR_EXTENSION_NAME)
 
-    val fatJarExtension = project.createKtorExtension<FatJarExtension>(FAT_JAR_EXTENSION_NAME)
+    // Apply Shadow plugin only when the application plugin is applied.
+    // TODO: KMP support will be added in Shadow 9.0.0
+    //   https://github.com/GradleUp/shadow/pull/1333
+    pluginManager.withPlugin(ApplicationPlugin.APPLICATION_PLUGIN_NAME) {
+        apply<ShadowPlugin>()
+    }
+
+    // By using `withPlugin` we handle the case when a user explicitly applies Shadow plugin.
+    pluginManager.withPlugin(SHADOW_PLUGIN_ID) {
+        configureShadowPlugin(fatJarExtension)
+    }
+}
+
+private fun Project.configureShadowPlugin(fatJarExtension: FatJarExtension) {
     val shadowJar: TaskProvider<ShadowJar> = tasks.named(SHADOW_JAR_TASK_NAME, ShadowJar::class.java) {
         it.archiveFileName.set(fatJarExtension.archiveFileName)
         it.isZip64 = fatJarExtension.allowZip64.get()