+++ plugin/src/test/kotlin/io/ktor/plugin/GradleVersionCompatibilityTest.kt
@@ -1,17 +1,12 @@
 package io.ktor.plugin
 
-import org.junit.jupiter.api.io.TempDir
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
-import java.io.File
 
-class GradleVersionCompatibilityTest {
+class GradleVersionCompatibilityTest : IntegrationTest() {
     @ParameterizedTest
-    @ValueSource(strings = ["8.3", "8.10.2"])
-    fun testProjectBuild(gradleVersion: String, @TempDir projectDir: File) {
-        projectDir.resolve("build.gradle.kts").writeText("""plugins { id("io.ktor.plugin") }""")
-        createGradleRunner(projectDir)
-            .withGradleVersion(gradleVersion)
-            .build()
+    @ValueSource(strings = ["8.3", "8.14"])
+    fun testProjectBuild(gradleVersion: String) {
+        runBuild { withGradleVersion(gradleVersion) }
     }
 }