+++ plugin/src/test/kotlin/io/ktor/plugin/TestConstants.kt
@@ -1,0 +1,3 @@
+package io.ktor.plugin
+
+internal const val KOTLIN_VERSION = "2.1.20"