+++ CHANGELOG.md
@@ -4,7 +4,16 @@ The format is based on Keep a Changelog: https://keepachangelog.com/en/1.1.0/
 
 ## [Unreleased]
 
-_No changes yet_
+- [KTOR-8444] Basic compatibility with the Kotlin Multiplatform Gradle plugin
+- [KTOR-8419] Do not apply Gradle's Application plugin for KMP projects
+
+[KTOR-8444]: https://youtrack.jetbrains.com/issue/KTOR-8444/
+[KTOR-8419]: https://youtrack.jetbrains.com/issue/KTOR-8419/
+
+## [3.1.3] - 2025-05-05
+
+- Update Ktor to [v3.1.3](https://github.com/ktorio/ktor/releases/tag/3.1.3)
+- Update Gradle to v8.14
 
 ## [3.1.2] - 2025-03-27
 
@@ -47,7 +56,8 @@ _No changes yet_
 - Bump default JRE in Docker to 21 (current LTS version)
 - Set Kotlin API and language level to 1.8 for compatibility with Gradle 8.0+
 
-[unreleased]: https://github.com/ktorio/ktor-build-plugins/compare/v3.1.2...main
+[unreleased]: https://github.com/ktorio/ktor-build-plugins/compare/v3.1.3...main
+[3.1.3]: https://github.com/ktorio/ktor-build-plugins/compare/v3.1.2...v3.1.3
 [3.1.2]: https://github.com/ktorio/ktor-build-plugins/compare/v3.1.1...v3.1.2
 [3.1.1]: https://github.com/ktorio/ktor-build-plugins/compare/v3.1.0...v3.1.1
 [3.1.0]: https://github.com/ktorio/ktor-build-plugins/compare/v3.0.3...v3.1.0